All the bugs are fixed and code is working ✅
